+++ lib/app_profiler.rb
@@ -10,6 +10,10 @@ module AppProfiler
   class ConfigurationError < StandardError
   end
 
+  DefaultProfileFormatter = proc do |upload|
+    "#{AppProfiler.speedscope_host}#profileURL=#{upload.url}"
+  end
+
   module Storage
     autoload :BaseStorage, "app_profiler/storage/base_storage"
     autoload :FileStorage, "app_profiler/storage/file_storage"
@@ -35,7 +39,8 @@ module AppProfiler
   mattr_accessor :autoredirect, default: false
   mattr_reader   :profile_header, default: "X-Profile"
   mattr_accessor :context, default: nil
-  mattr_reader   :profile_url_formatter, default: nil
+  mattr_reader   :profile_url_formatter,
+    default: DefaultProfileFormatter
 
   mattr_accessor :storage, default: Storage::FileStorage
   mattr_accessor :viewer, default: Viewer::SpeedscopeViewer
@@ -71,6 +76,10 @@ module AppProfiler
 
     def profile_url_formatter=(block)
       @@profile_url_formatter = block # rubocop:disable Style/ClassVars
+    end
+
+    def profile_url(upload)
+      AppProfiler.profile_url_formatter.call(upload)
     end
   end
 end