+++ lib/app_profiler/profile.rb
@@ -40,7 +40,13 @@ module AppProfiler
     def upload
       AppProfiler.storage.upload(self).tap do |upload|
         if upload && defined?(upload.url)
-          AppProfiler.logger.info("[Profiler] uploaded profile profile_url=#{upload.url}")
+          AppProfiler.logger.info(
+            <<~INFO.squish
+              [Profiler] data uploaded:
+              profile_url=#{upload.url}
+              profile_viewer_url=#{AppProfiler.profile_url(upload)}
+            INFO
+          )
         end
       end
     rescue => error