+++ test/support/test_helper.rb
@@ -15,7 +15,9 @@ module AppProfiler
 
     def assert_profiles_uploaded(autoredirect: false)
       AppProfiler.stubs(:storage).returns(MockStorage)
-      AppProfiler.logger.expects(:info).with { |value| value =~ /uploaded profile/ }
+      AppProfiler.logger.expects(:info).with do |value|
+        value =~ /\[Profiler\] data uploaded: profile_url=.* profile_viewer_url=.*/
+      end
 
       response = yield